refactored the backend routes folder
