+++ endpoints/index.js
@@ -4,18 +4,9 @@ const router = express.Router();
 const cors = require("cors");
 
 const auth = require("./auth");
-const createTournament = require("./admin/createTournament");
-const tournamentCreateBrackets = require("./admin/tournamentCreateBrackets");
-const tournamentWeightClasses = require("./admin/tournamentWeightClasses");
 const admin = require("./admin");
-
-const tournamentSignUp = require("./tournamentSignUp");
-const tournamentBrackets = require("./tournamentBrackets");
-const tournamentMatchResults = require("./tournamentMatchResults");
-const getTournamentCardData = require("./getTournamentCardData");
-const searchTournament = require("./searchTournament");
-const getTournamentSignUps = require("./getTournamentSignUps");
-const getTournamentBracketData = require("./getTournamentBracketData");
+const public = require("./public");
+const wrestler = require("./wrestler");
 
 router.use(cors());
 
@@ -43,12 +34,12 @@ router.use(
   isAdmin,
   admin.tournamentCreateBrackets
 );
-router.use("/tournamentSignUp", isWrestler, tournamentSignUp);
-router.use("/tournamentBrackets", tournamentBrackets);
-router.use("/tournamentMatchResults", tournamentMatchResults);
-router.use("/getTournamentCardData", getTournamentCardData);
-router.use("/searchTournament", searchTournament);
-router.use("/getTournamentSignUps", getTournamentSignUps);
-router.use("/getTournamentBracketData", getTournamentBracketData);
+router.use("/tournamentSignUp", isWrestler, wrestler.tournamentSignUp);
+router.use("/tournamentBrackets", public.tournamentBrackets);
+router.use("/tournamentMatchResults", public.tournamentMatchResults);
+router.use("/getTournamentCardData", public.tournamentCardData);
+router.use("/searchTournament", public.searchTournament);
+router.use("/getTournamentSignUps", public.getTournamentSignUps);
+router.use("/tournamentBracketData", public.tournamentBracketData);
 
 module.exports = router;
