+++ endpoints/public/getTournamentSignUps.js
