+++ endpoints/public/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  searchTournament: require("./searchTournament"),
+  tournamentBracketData: require("./tournamentBracketData"),
+  tournamentBrackets: require("./tournamentBrackets"),
+  tournamentCardData: require("./tournamentCardData"),
+  tournamentMatchResults: require("./tournamentMatchResults"),
+  tournamentSignUp: require("../wrestler/tournamentSignUp"),
+  getTournamentSignUps: require("./getTournamentSignUps")
+};
