+++ endpoints/public/searchTournament.js
