+++ endpoints/public/tournamentBrackets.js
