+++ endpoints/public/tournamentCardData.js
