+++ endpoints/public/tournamentMatchResults.js
