+++ endpoints/wrestler/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  tournamentSignUp: require("./tournamentSignUp")
+};
