+++ endpoints/wrestler/tournamentSignUp.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const tournament = require("../db/tournament");
-const search = require("../db/search");
+const tournament = require("../../db/tournament");
+const search = require("../../db/search");
 
 router.post("/", (req, res) => {
   const { seeding_notes, tournament_id, tournament_weight_class } = req.body;
